feat: validate if task exists before adding to data

+++ actions.py
@@ -1,21 +1,29 @@
 from pathlib import Path
 import json
-from functions import get_date_time
+from typing import List, TypedDict
+from functions import get_date_time, clear_terminal
 
-def add_task(task_name: str):
+# tipagem 
+class Task(TypedDict):
+    id: int
+    description: str
+    status: str
+    createAt: str  
+    updateAt: str
+
+
+def add_task(task_name: str) -> None:
     DIR_ROOT = Path(__file__).parent
     FILE_NAME = "data.json"
     FILE_PATH = DIR_ROOT / FILE_NAME
 
-    data = []
+    data: List[Task] = []
 
     if FILE_PATH.exists():
         try:
             with open(FILE_PATH, "r", encoding="utf-8") as f:
                 data = json.load(f)
-                # garante que data seja uma lista para podermos usar append
-                if not isinstance(data, list):
-                    data = []
+
         except (json.JSONDecodeError, ValueError):
             # Se o arquivo estiver corrompido ou vazio, começamos do zero
             data = []
@@ -25,12 +33,25 @@ def add_task(task_name: str):
     date_time = get_date_time()
 
     
-    new_task: dict[str, str | int] = {"id": len(data) + 1, "description": task_name, "status": "todo", "createAt": date_time , "updateAt": date_time } # type: ignore
+    new_task: Task = {"id": len(data) + 1, "description": task_name, "status": "todo", "createAt": date_time , "updateAt": date_time }
 
-    data.append(new_task) # type: ignore
+    # Verificando se a tarefa que o usuário quer inserir já existe
+    _exists = False
 
-    with open(FILE_PATH, "w", encoding="utf-8") as f:
-        json.dump(data, f, indent=4, ensure_ascii=False)
+    for d in data:
+        if task_name in d['description']:
+            _exists = True
+            break
     
-    print(f"Tarefa '{task_name}' adicionada com sucesso!")
+    if _exists:
+        clear_terminal()
+        print(f"\033[31m\nA terefa {task_name} já existe!\033[m")
     
+    else:
+        data.append(new_task) # type: ignore
+
+        with open(FILE_PATH, "w", encoding="utf-8") as f:
+            json.dump(data, f, indent=4, ensure_ascii=False)
+
+        clear_terminal() 
+        print(f"Tarefa '{task_name}' adicionada com sucesso!\n")
